Move test code from across `direct` to proper test modules
