+++ tests/stdpy/test_threading2.py
@@ -1,42 +1,16 @@
-import sys
+from collections import deque
 from panda3d import core
-from direct.stdpy import threading
-import pytest
+from direct.stdpy import threading2
 
 
-def test_threading_error():
-    with pytest.raises(threading.ThreadError):
-        threading.stack_size()
-
-
-def test_threading():
-    from collections import deque
-
-    _sleep = core.Thread.sleep
-
-    _VERBOSE = False
-
-    class _Verbose(object):
-
-        def __init__(self, verbose=None):
-            if verbose is None:
-                verbose = _VERBOSE
-            self.__verbose = verbose
-
-        def _note(self, format, *args):
-            if self.__verbose:
-                format = format % args
-                format = "%s: %s\n" % (
-                    threading.currentThread().getName(), format)
-                sys.stderr.write(format)
-
-    class BoundedQueue(_Verbose):
+def test_threading2():
+    class BoundedQueue(threading2._Verbose):
 
         def __init__(self, limit):
-            _Verbose.__init__(self)
-            self.mon = threading.Lock(name = "BoundedQueue.mon")
-            self.rc = threading.Condition(self.mon)
-            self.wc = threading.Condition(self.mon)
+            threading2._Verbose.__init__(self)
+            self.mon = threading2.RLock()
+            self.rc = threading2.Condition(self.mon)
+            self.wc = threading2.Condition(self.mon)
             self.limit = limit
             self.queue = deque()
 
@@ -62,10 +36,10 @@ def test_threading():
             self.mon.release()
             return item
 
-    class ProducerThread(threading.Thread):
+    class ProducerThread(threading2.Thread):
 
         def __init__(self, queue, quota):
-            threading.Thread.__init__(self, name="Producer")
+            threading2.Thread.__init__(self, name="Producer")
             self.queue = queue
             self.quota = quota
 
@@ -75,12 +49,12 @@ def test_threading():
             while counter < self.quota:
                 counter = counter + 1
                 self.queue.put("%s.%d" % (self.getName(), counter))
-                _sleep(random() * 0.00001)
+                core.Thread.sleep(random() * 0.00001)
 
-    class ConsumerThread(threading.Thread):
+    class ConsumerThread(threading2.Thread):
 
         def __init__(self, queue, count):
-            threading.Thread.__init__(self, name="Consumer")
+            threading2.Thread.__init__(self, name="Consumer")
             self.queue = queue
             self.count = count
 
@@ -103,7 +77,7 @@ def test_threading():
     C = ConsumerThread(Q, NI*NP)
     for t in P:
         t.start()
-        _sleep(0.000001)
+        core.Thread.sleep(0.000001)
     C.start()
     for t in P:
         t.join()
