+++ tests/tkwidgets/test_AppShell.py
@@ -1,0 +1,5 @@
+from direct.tkwidgets import AppShell
+
+
+def test_TestAppShell():
+    test = AppShell.TestAppShell(balloon_state='none')
